Sprint four

+++ src/main/java/ru/ythree/blog/repository/JdbcNativePostRepository.java
@@ -34,7 +34,7 @@ public class JdbcNativePostRepository implements PostRepository {
     }
 
     @Override
-    public List<Post> findAll(SearchFilter searchFilter, int offset, int size) {
+    public List<Post> findAll(SearchFilter searchFilter, Integer offset, Integer size) {
         MapSqlParameterSource params = new MapSqlParameterSource();
         params.addValue("limit", size);
         params.addValue("offset", offset);
@@ -163,8 +163,8 @@ public class JdbcNativePostRepository implements PostRepository {
                         rs.getString("title"),
                         rs.getString("text"),
                         tags == null ? new String[]{} : tags.split(","),
-                        rs.getInt("likesCount"),
-                        rs.getInt("commentsCount"));
+                        rs.getLong("likesCount"),
+                        rs.getLong("commentsCount"));
                 posts.put(postId, post);
             }
         };
